Data type changes and comments
Added comments and changed data types to doubles instead of int

+++ System/AerodynamicsSimulationAdapter.cpp
@@ -10,7 +10,7 @@ AerodynamicsSimulationAdapter::~AerodynamicsSimulationAdapter(){
 }
 
 void AerodynamicsSimulationAdapter::calculateLevel(){
-	int l = 0;
+	double l = 0;
 
 	double widthCalc = 1 / (part->getFrontWingWidth() - part->getBackWingWidth());
